requirements.txt
azure-functions
requests
cryptography>=43.0.0  # Latest version as of April 2025
azure-identity
azure-mgmt-dns
azure-keyvault-certificates
azure-keyvault-secrets
